+++ draft-packages/card/docs/Card.stories.tsx
@@ -3,9 +3,10 @@ import * as React from "react"
 import { withDesign } from "storybook-addon-designs"
 import { Card } from ".."
 import { figmaEmbed } from "../../../storybook/helpers"
+import { CATEGORIES } from "../../../storybook/constants"
 
 export default {
-  title: "Card (React)",
+  title: `${CATEGORIES.components}/Card`,
   component: Card,
   parameters: {
     docs: {
@@ -40,5 +41,5 @@ export const CardWithSpace = () => (
 CardWithSpace.storyName = "Card, custom spacing with Box"
 
 CardWithSpace.parameters = {
-  backgrounds: { default: "Stone" },
+  backgrounds: { default: "Gray 100" },
 }