+++ draft-packages/form/docs/ElmTextAreaField.stories.tsx
@@ -1,9 +1,10 @@
 import { loadElmStories } from "elm-storybook"
+import { CATEGORIES } from "../../../storybook/constants"
 
 const compiledElm = require("../ElmStories/TextAreaFieldStories.elm").Elm
   .ElmStories.TextAreaFieldStories
 
-loadElmStories("TextAreaField (Elm)", module, compiledElm, [
+loadElmStories(`${CATEGORIES.elm}/Text Area Field`, module, compiledElm, [
   "Default",
   "Default, Controlled, Prefilled Value",
   "Default, Uncontrolled, Prefilled Value",