+++ draft-packages/illustration/KaizenDraft/Illustration/Players/VideoPlayer.spec.tsx
@@ -39,7 +39,7 @@ describe("<VideoPlayer />", () => {
       <VideoPlayer
         autoplay
         fallback="illustrations/heart/spot/moods-cautionary.svg"
-        ambientAnimation="illustrations/heart/spot/moods-cautionary.webm"
+        source="illustrations/heart/spot/moods-cautionary.webm"
       />
     )
     const videoPlayer = screen.getByTestId("kz-video-player")
@@ -56,7 +56,7 @@ describe("<VideoPlayer />", () => {
         <VideoPlayer
           autoplay
           fallback="illustrations/heart/spot/moods-cautionary.svg"
-          ambientAnimation="illustrations/heart/spot/moods-cautionary.webm"
+          source="illustrations/heart/spot/moods-cautionary.webm"
         />
       )
       const videoPlayer = screen.getByTestId("kz-video-player")
@@ -71,10 +71,6 @@ describe("<VideoPlayer />", () => {
           preload="metadata"
           width="100%"
         >
-          <source
-            src="https://d1e7r7b0lb8p4d.cloudfront.net/illustrations/heart/spot/moods-cautionary.webm.webm"
-            type="video/webm"
-          />
           <source
             src="https://d1e7r7b0lb8p4d.cloudfront.net/illustrations/heart/spot/moods-cautionary.webm.mp4"
             type="video/mp4"
@@ -92,7 +88,7 @@ describe("<VideoPlayer />", () => {
         <VideoPlayer
           autoplay
           fallback="illustrations/heart/spot/moods-cautionary.svg"
-          ambientAnimation="illustrations/heart/spot/moods-cautionary.webm"
+          source="illustrations/heart/spot/moods-cautionary.webm"
         />
       )
       const videoPlayer = screen.getByTestId("kz-video-player")
@@ -108,10 +104,6 @@ describe("<VideoPlayer />", () => {
           preload="metadata"
           width="100%"
         >
-          <source
-            src="https://d1e7r7b0lb8p4d.cloudfront.net/illustrations/heart/spot/moods-cautionary.webm.webm"
-            type="video/webm"
-          />
           <source
             src="https://d1e7r7b0lb8p4d.cloudfront.net/illustrations/heart/spot/moods-cautionary.webm.mp4"
             type="video/mp4"