+++ draft-packages/loading-placeholder/docs/LoadingPlaceholder.stories.tsx
@@ -2,6 +2,7 @@ import { Heading, Paragraph } from "@kaizen/component-library"
 import { LoadingPlaceholder } from "@kaizen/draft-loading-placeholder"
 import * as React from "react"
 import { withDesign } from "storybook-addon-designs"
+import { CATEGORIES } from "../../../storybook/constants"
 import { figmaEmbed } from "../../../storybook/helpers"
 
 import styles from "./LoadingPlaceholder.stories.scss"
@@ -11,7 +12,7 @@ const StoryContainer: React.FunctionComponent = ({ children }) => (
 )
 
 export default {
-  title: "LoadingPlaceholder (React)",
+  title: `${CATEGORIES.components}/Loading Placeholder`,
   component: LoadingPlaceholder,
   parameters: {
     docs: {
@@ -248,7 +249,7 @@ export const ReversedDefault = () => (
 ReversedDefault.storyName = "Reversed, Default"
 ReversedDefault.parameters = {
   backgrounds: {
-    default: "Wisteria 700",
+    default: "Purple 700",
   },
 }