+++ draft-packages/menu/docs/Menu.stories.tsx
@@ -19,6 +19,7 @@ import {
   MenuSeparator,
   StatelessMenu,
 } from ".."
+import { CATEGORIES } from "../../../storybook/constants"
 
 const MenuInstance: React.FunctionComponent = () => (
   <MenuContent>
@@ -27,6 +28,14 @@ const MenuInstance: React.FunctionComponent = () => (
     <MenuItem action="https://www.cultureamp.com/" label="Menu link" />
     <MenuSeparator />
     <MenuHeader title="Other actions" />
+    <MenuItem
+      action={(e: any) => {
+        alert("Hello")
+        e.preventDefault()
+      }}
+      icon={editIcon}
+      label="Menu button"
+    />
     <MenuItem
       action={(e: any) => {
         e.preventDefault()
@@ -38,26 +47,45 @@ const MenuInstance: React.FunctionComponent = () => (
       action={(e: any) => {
         e.preventDefault()
       }}
-      icon={editIcon}
-      label="Menu button"
-    />
-
-    <MenuItem
-      action="https://www.cultureamp.com/"
       icon={trashIcon}
       destructive
-      label="Menu button"
+      label="Destructive Menu button"
     />
     <MenuItem
-      action="https://www.cultureamp.com/"
+      action={(e: any) => {
+        e.preventDefault()
+      }}
+      icon={trashIcon}
       disabled
-      label="Menu button"
+      label="Disabled Menu button"
+    />
+    <MenuItem
+      action={(e: any) => {
+        e.preventDefault()
+      }}
+      icon={trashIcon}
+      disabled
+      destructive
+      label="Disabled Destructive Menu button"
+    />
+    <MenuItem
+      action={(e: any) => {
+        e.preventDefault()
+      }}
+      label="Menu button no icon"
+    />
+    <MenuItem
+      action={(e: any) => {
+        e.preventDefault()
+      }}
+      disabled
+      label="Disabled button no icon"
     />
   </MenuContent>
 )
 
 export default {
-  title: "Menu (React)",
+  title: `${CATEGORIES.components}/Menu`,
   component: Menu,
   parameters: {
     docs: {
@@ -138,7 +166,7 @@ export const LabelAndIconReversed = () => (
 LabelAndIconReversed.storyName = "Label and Icon (reversed)"
 LabelAndIconReversed.parameters = {
   backgrounds: {
-    default: "Wisteria 700",
+    default: "Purple 700",
   },
 }