+++ draft-packages/tabs/docs/Tabs.stories.tsx
@@ -6,10 +6,11 @@ import classnames from "classnames"
 import * as React from "react"
 import { withDesign } from "storybook-addon-designs"
 import { figmaEmbed } from "../../../storybook/helpers"
+import { CATEGORIES } from "../../../storybook/constants"
 import { ExampleLayout } from "./ExampleLayout"
 
 export default {
-  title: "Tabs (React)",
+  title: `${CATEGORIES.components}/Tabs`,
   component: Tabs,
   parameters: {
     docs: {
@@ -311,7 +312,7 @@ export const WithLayoutVerticalLTR = () => {
 WithLayoutVerticalLTR.storyName = "(Example) Layout LTR (Vertical)"
 
 WithLayoutVerticalLTR.parameters = {
-  backgrounds: { default: "Wisteria 700" },
+  backgrounds: { default: "Purple 700" },
 }
 
 export const WithLayoutVerticalRTL = () => {
@@ -342,7 +343,7 @@ export const WithLayoutVerticalRTL = () => {
 WithLayoutVerticalRTL.storyName = "(Example) Layout RTL (Vertical)"
 
 WithLayoutVerticalRTL.parameters = {
-  backgrounds: { default: "Stone" },
+  backgrounds: { default: "Gray 100" },
 }
 
 export const WithHeading = () => {
@@ -386,7 +387,7 @@ export const WithHeading = () => {
 WithHeading.storyName = "(Example) Layout With heading"
 
 WithHeading.parameters = {
-  backgrounds: { default: "Stone" },
+  backgrounds: { default: "Gray 100" },
 }
 
 export const ExampleContentTab = () => {
@@ -420,5 +421,5 @@ export const ExampleContentTab = () => {
 ExampleContentTab.storyName = "(Example) Content Tab in Content Area"
 
 ExampleContentTab.parameters = {
-  backgrounds: { default: "Stone" },
+  backgrounds: { default: "Gray 100" },
 }