+++ draft-packages/tooltip/docs/Tooltip.stories.tsx
@@ -9,19 +9,20 @@ import { Button } from "@kaizen/draft-button"
 import { Tooltip } from "@kaizen/draft-tooltip"
 import isChromatic from "chromatic/isChromatic"
 import { figmaEmbed } from "../../../storybook/helpers"
+import { CATEGORIES } from "../../../storybook/constants"
 
 /**
  * We should not be running visual regressions on these tooltip stories
  * until there is a clean way to trigger open states for the snapshots. See:
  * https://github.com/cultureamp/kaizen-design-system/issues/1375
  */
-const excludeFromChromatic = Story => {
-  if (!isChromatic()) return <Story />
-  return null
+const openTooltipInChromatic = (story, config) => {
+  if (isChromatic()) config.args.isInitiallyVisible = true
+  return story()
 }
 
 export default {
-  title: "Tooltip (React)",
+  title: `${CATEGORIES.components}/Tooltip`,
   component: Tooltip,
   parameters: {
     docs: {
@@ -33,12 +34,12 @@ export default {
       "https://www.figma.com/file/GMxm8rvDCbj0Xw3TQWBZ8b/UI-Kit-Zen?node-id=14473%3A90872"
     ),
   },
-  decorators: [withDesign, excludeFromChromatic],
+  decorators: [withDesign, openTooltipInChromatic],
 }
 
-export const DefaultBelowKaizenSiteDemo = () => (
+export const DefaultBelowKaizenSiteDemo = props => (
   <div style={{ display: "flex", justifyContent: "center" }}>
-    <Tooltip position="below" text="This is below the tooltip">
+    <Tooltip {...props} position="below" text="This is below the tooltip">
       {/* Using buttons, as so we can test the focus state.
          ie. the tooltip should show when any child is focused. */}
       <Button label="Below" />
@@ -47,8 +48,6 @@ export const DefaultBelowKaizenSiteDemo = () => (
 )
 
 DefaultBelowKaizenSiteDemo.storyName = "Default - Below (Kaizen Site Demo)"
-DefaultBelowKaizenSiteDemo.component = Tooltip
-
 DefaultBelowKaizenSiteDemo.parameters = {
   info: {
     text: `
@@ -57,12 +56,12 @@ DefaultBelowKaizenSiteDemo.parameters = {
   },
 }
 
-export const DefaultAbove = () => (
+export const DefaultAbove = props => (
   <div
     style={{ marginTop: "100px", display: "flex", justifyContent: "center" }}
   >
     <div style={{ display: "inline-block", position: "relative" }}>
-      <Tooltip position="above" text="This is above the tooltip">
+      <Tooltip {...props} position="above" text="This is above the tooltip">
         {/* Using buttons, as so we can test the focus state.
          ie. the tooltip should show when any child is focused. */}
         <Button label="Above" />
@@ -70,10 +69,9 @@ export const DefaultAbove = () => (
     </div>
   </div>
 )
-
 DefaultAbove.storyName = "Default - Above"
 
-export const Inline = () => (
+export const Inline = props => (
   <div
     style={{
       margin: "100px",
@@ -86,12 +84,16 @@ export const Inline = () => (
       <Paragraph variant="body">
         Lorem ipsum dolor sit amet consectetur adipisicing elit. Ratione nulla
         quas corporis? Perspiciatis, ratione voluptas{" "}
-        <Tooltip display="inline-block" text="This is above the tooltip">
+        <Tooltip
+          {...props}
+          display="inline-block"
+          text="This is above the tooltip"
+        >
           <Tag>ad veniam sapiente</Tag>
         </Tooltip>{" "}
         Maxime harum, ducimus maiores itaque pariatur quod vel porro mollitia.
         Lorem ipsum dolor sit{" "}
-        <Tooltip display="inline" text="Open in new tab">
+        <Tooltip {...props} display="inline" text="Open in new tab">
           <a href="#">
             amet consectetur adipisicing elit Itaque obcaecati maxime molestiae
             blanditiis pariatur
@@ -100,6 +102,7 @@ export const Inline = () => (
         . Magni perspiciatis assumenda in adipisci, eaque commodi quidem dolore,
         tempore provident animi{" "}
         <Tooltip
+          {...props}
           display="inline-block"
           text="This is below the tooltip"
           position="below"
@@ -113,10 +116,11 @@ export const Inline = () => (
 
 Inline.storyName = "Inline"
 
-export const ArrowPositioning = () => (
+export const ArrowPositioning = props => (
   <div>
     <div style={{ position: "absolute", top: 0, left: 0 }}>
       <Tooltip
+        {...props}
         position="above"
         text="This is below the tooltip, despite it being set to position=above. This is because there is not enough room. Also note that the arrow is correctly positioned."
       >
@@ -125,6 +129,7 @@ export const ArrowPositioning = () => (
     </div>
     <div style={{ position: "absolute", bottom: 0, right: 0 }}>
       <Tooltip
+        {...props}
         position="below"
         text="This is above the tooltip, despite it being set to position=below. This is because there is not enough room. Also note that the arrow is correctly positioned."
       >
@@ -136,11 +141,12 @@ export const ArrowPositioning = () => (
 
 ArrowPositioning.storyName = "Arrow positioning"
 
-export const OverflowScroll = () => (
+export const OverflowScroll = props => (
   <div>
     <div style={{ overflowX: "scroll", width: "200px", height: "100px" }}>
       <div style={{ width: "500px", textAlign: "center" }}>
         <Tooltip
+          {...props}
           position="below"
           display="inline-block"
           text="This should not get cropped"
@@ -162,7 +168,7 @@ export const OverflowScroll = () => (
 
 OverflowScroll.storyName = "overflow: scroll"
 
-export const TooltipAboveDropdown = () => (
+export const TooltipAboveDropdown = props => (
   <>
     <div>
       <SplitButton
@@ -180,6 +186,7 @@ export const TooltipAboveDropdown = () => (
     </div>
     <div style={{ paddingTop: "50px" }}>
       <Tooltip
+        {...props}
         position="above"
         display="inline-block"
         text="Text to appear above the button dropdown text"