+++ packages/component-library/components/Text/Text.tsx
@@ -3,6 +3,7 @@
 import classNames from "classnames"
 import * as React from "react"
 
+import { withDeprecatedComponent } from "@kaizen/react-deprecate-warning"
 import styles from "./Text.module.scss"
 
 type TextProps = {
@@ -60,4 +61,6 @@ const Text: React.FunctionComponent<TextProps> = ({
 
 Text.displayName = "Text"
 
-export default Text
+export default withDeprecatedComponent(Text, {
+  warning: "Text is deprecated. Use Paragraph or Heading instead.",
+})