+++ draft-packages/popover/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-popover@2.2.10...@kaizen/draft-popover@2.2.11) (2021-08-02)
+
+**Note:** Version bump only for package @kaizen/draft-popover
+
+
+
+
+
+## [2.2.10](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-popover@2.2.9...@kaizen/draft-popover@2.2.10) (2021-08-02)
+
+**Note:** Version bump only for package @kaizen/draft-popover
+
+
+
+
+
+## [2.2.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-popover@2.2.8...@kaizen/draft-popover@2.2.9) (2021-07-30)
+
+**Note:** Version bump only for package @kaizen/draft-popover
+
+
+
+
+
 ## [2.2.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-popover@2.2.7...@kaizen/draft-popover@2.2.8) (2021-07-28)
 
 **Note:** Version bump only for package @kaizen/draft-popover