+++ draft-packages/slider/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-slider@1.4.10...@kaizen/draft-slider@1.4.11) (2021-08-02)
+
+**Note:** Version bump only for package @kaizen/draft-slider
+
+
+
+
+
+## [1.4.10](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-slider@1.4.9...@kaizen/draft-slider@1.4.10) (2021-08-02)
+
+**Note:** Version bump only for package @kaizen/draft-slider
+
+
+
+
+
+## [1.4.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-slider@1.4.8...@kaizen/draft-slider@1.4.9) (2021-07-30)
+
+**Note:** Version bump only for package @kaizen/draft-slider
+
+
+
+
+
 ## [1.4.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-slider@1.4.7...@kaizen/draft-slider@1.4.8) (2021-07-28)
 
 **Note:** Version bump only for package @kaizen/draft-slider