+++ draft-packages/well/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.27](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@3.1.26...@kaizen/draft-well@3.1.27) (2021-08-02)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [3.1.26](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@3.1.25...@kaizen/draft-well@3.1.26) (2021-08-02)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [3.1.25](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@3.1.24...@kaizen/draft-well@3.1.25) (2021-07-30)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
 ## [3.1.24](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@3.1.23...@kaizen/draft-well@3.1.24) (2021-07-28)
 
 **Note:** Version bump only for package @kaizen/draft-well