+++ legacy-packages/title-block/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.21](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@3.1.20...@kaizen/draft-title-block@3.1.21) (2021-08-02)
+
+**Note:** Version bump only for package @kaizen/draft-title-block
+
+
+
+
+
+## [3.1.20](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@3.1.19...@kaizen/draft-title-block@3.1.20) (2021-08-02)
+
+
+### Bug Fixes
+
+* Add deprecated jsdocs and wrap with withDeprecatedComponent ([#1787](https://github.com/cultureamp/kaizen-design-system/issues/1787)) ([626188e](https://github.com/cultureamp/kaizen-design-system/commit/626188e15f1b7ff7bf2046793c94e7b569fc3f1b))
+
+
+
+
+
+## [3.1.19](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@3.1.18...@kaizen/draft-title-block@3.1.19) (2021-07-30)
+
+**Note:** Version bump only for package @kaizen/draft-title-block
+
+
+
+
+
+## [3.1.18](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@3.1.17...@kaizen/draft-title-block@3.1.18) (2021-07-29)
+
+**Note:** Version bump only for package @kaizen/draft-title-block
+
+
+
+
+
 ## [3.1.17](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@3.1.16...@kaizen/draft-title-block@3.1.17) (2021-07-28)
 
 **Note:** Version bump only for package @kaizen/draft-title-block