+++ packages/brand/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand@1.1.2...@kaizen/brand@1.1.3) (2021-08-02)
+
+**Note:** Version bump only for package @kaizen/brand
+
+
+
+
+
+## [1.1.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand@1.1.1...@kaizen/brand@1.1.2) (2021-07-30)
+
+
+### Bug Fixes
+
+* Update Storybook background colour references ([#1803](https://github.com/cultureamp/kaizen-design-system/issues/1803)) ([118f242](https://github.com/cultureamp/kaizen-design-system/commit/118f24201133aa5fd42839b67ad7cd74273d02e9))
+
+
+
+
+
 ## [1.1.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand@1.1.0...@kaizen/brand@1.1.1) (2021-07-15)